+++ aiogram_dialog/api/entities/__init__.py
@@ -1,6 +1,7 @@
 __all__ = [
     "Context", "Data",
     "ChatEvent",
+    "LaunchMode",
     "MediaAttachment", "MediaId",
     "ShowMode", "StartMode",
     "DEFAULT_STACK_ID", "Stack",
@@ -10,6 +11,7 @@ __all__ = [
 
 from .context import Context, Data
 from .events import ChatEvent
+from .launch_mode import LaunchMode
 from .media import MediaAttachment, MediaId
 from .modes import ShowMode, StartMode
 from .stack import DEFAULT_STACK_ID, Stack
